time_series/requirements.txt
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
torch>=2.1.0
matplotlib>=3.7.0
PyYAML>=6.0
statsmodels>=0.13.0  # optional: ARIMA
